several logs - fixed context menu and prioritized filter
