+++ Src/WitsmlExplorer.Frontend/components/ContextMenus/LogCurveInfoContextMenu.tsx
@@ -1,7 +1,7 @@
 import { Divider, Typography } from "@equinor/eds-core-react";
 import { MenuItem } from "@mui/material";
 import { WITSML_INDEX_TYPE_MD } from "components/Constants";
-import { LogCurveInfoRow } from "components/ContentViews/LogCurveInfoListView";
+import { LogCurveInfoRow } from "components/ContentViews/LogCurveInfoListViewUtils";
 import ContextMenu from "components/ContextMenus/ContextMenu";
 import {
   StyledIcon,
